+++ src/Backend/Http/Controllers/FileManager/DeleteController.php
@@ -3,8 +3,8 @@
 namespace Mymo\Backend\Http\Controllers\FileManager;
 
 use Illuminate\Http\Request;
-use Mymo\Core\Models\Files;
-use Mymo\Core\Models\Folders;
+use Mymo\Core\Models\File;
+use Mymo\Core\Models\Folder;
 
 class DeleteController extends FileManagerController
 {
@@ -21,10 +21,10 @@ class DeleteController extends FileManagerController
     
             $is_directory = $this->isDirectory($file);
             if ($is_directory) {
-                Folders::find($file)->deleteFolder();
+                Folder::find($file)->deleteFolder();
             } else {
                 $file_path = $this->getPath($file);
-                Files::where('path', '=', $file_path)
+                File::where('path', '=', $file_path)
                     ->first()
                     ->delete();
             }