+++ src/Backend/Http/Controllers/FileManager/DownloadController.php
@@ -3,14 +3,14 @@
 namespace Mymo\Backend\Http\Controllers\FileManager;
 
 use Illuminate\Support\Facades\Storage;
-use Mymo\Core\Models\Files;
+use Mymo\Core\Models\File;
 
 class DownloadController extends FileManagerController
 {
     public function getDownload()
     {
         $file = $this->getPath(request()->get('file'));
-        $data = Files::where('path', '=', $file)->first(['name']);
+        $data = File::where('path', '=', $file)->first(['name']);
         
         $path = Storage::disk(config('mymo.filemanager.disk'))->path($file);
         if ($data) {