+++ src/Backend/Http/Controllers/FileManager/ItemsController.php
@@ -3,8 +3,8 @@
 namespace Mymo\Backend\Http\Controllers\FileManager;
 
 use Illuminate\Support\Facades\Storage;
-use Mymo\Core\Models\Files;
-use Mymo\Core\Models\Folders;
+use Mymo\Core\Models\File;
+use Mymo\Core\Models\Folder;
 
 class ItemsController extends FileManagerController
 {
@@ -16,11 +16,11 @@ class ItemsController extends FileManagerController
         
         $working_dir = request()->get('working_dir');
         
-        $folders = Folders::where('folder_id', '=', $working_dir)
+        $folders = Folder::where('folder_id', '=', $working_dir)
             ->where('type', '=', $file_type)
             ->orderBy('name', 'ASC')
             ->get(['id', 'name']);
-        $files = Files::where('folder_id', '=', $working_dir)
+        $files = File::where('folder_id', '=', $working_dir)
             ->where('type', '=', $file_type)
             ->orderBy('id', 'DESC')
             ->paginate($perPage);