+++ src/Backend/Http/Controllers/FileManager/RenameController.php
@@ -2,8 +2,8 @@
 
 namespace Mymo\Backend\Http\Controllers\FileManager;
 
-use Mymo\Core\Models\Files;
-use Mymo\Core\Models\Folders;
+use Mymo\Core\Models\File;
+use Mymo\Core\Models\Folder;
 
 class RenameController extends FileManagerController
 {
@@ -23,14 +23,14 @@ class RenameController extends FileManagerController
         }
 
         if ($is_directory) {
-            Folders::where('id', '=', $file)
+            Folder::where('id', '=', $file)
                 ->update([
                     'name' => $new_name
                 ]);
         } else {
             $file_path = explode('uploads/', $file)[1];
             
-            Files::where('path', '=', $file_path)
+            File::where('path', '=', $file_path)
                 ->update([
                     'name' => $new_name
                 ]);